// src/distribution/models/transmittalItem.types.ts
export enum ChoicesStatusTransmittalItem {
  pending = 'pending',
  shipped = 'shipped',
  loaded = 'loaded',
  documented = 'documented',
  arrived = 'arrived',
  received = 'received',
  damaged = 'damaged',
  completed = 'completed'
}

export const choicesStatusTransmittalItem = [
  ChoicesStatusTransmittalItem.pending,
  ChoicesStatusTransmittalItem.shipped,
  ChoicesStatusTransmittalItem.loaded,
  ChoicesStatusTransmittalItem.documented,
  ChoicesStatusTransmittalItem.arrived,
  ChoicesStatusTransmittalItem.received,
  ChoicesStatusTransmittalItem.damaged,
  ChoicesStatusTransmittalItem.completed,
];